requirements.txt
flask>=2.0.0
flask-socketio>=5.3.0
python-socketio>=5.0.0
gunicorn>=20.1.0
gevent>=23.0.0
gevent-websocket>=0.10.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
reportlab>=4.0.0
boto3>=1.28.0
